requirements.txt
# Django and Web Server
django

# Machine Learning & OCR
ultralytics
easyocr
langdetect
opencv-python-headless
Pillow
numpy
torch
torchvision
torchaudio
